+++ app/pages/project/access/ProjectAccessRolesTab.tsx
@@ -1,0 +1,234 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, you can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * Copyright Oxide Computer Company
+ */
+import { createColumnHelper, getCoreRowModel, useReactTable } from '@tanstack/react-table'
+import { useMemo, useState } from 'react'
+import type { LoaderFunctionArgs } from 'react-router'
+import * as R from 'remeda'
+
+import {
+  api,
+  byGroupThenName,
+  deleteRole,
+  q,
+  queryClient,
+  roleOrder,
+  useApiMutation,
+  usePrefetchedQuery,
+  useUserRows,
+  type IdentityType,
+  type RoleKey,
+} from '@oxide/api'
+import { Access24Icon } from '@oxide/design-system/icons/react'
+import { Badge } from '@oxide/design-system/ui'
+
+import { HL } from '~/components/HL'
+import { ListPlusCell } from '~/components/ListPlusCell'
+import {
+  ProjectAccessAddUserSideModal,
+  ProjectAccessEditUserSideModal,
+} from '~/forms/project-access'
+import { titleCrumb } from '~/hooks/use-crumbs'
+import { getProjectSelector, useProjectSelector } from '~/hooks/use-params'
+import { confirmDelete } from '~/stores/confirm-delete'
+import { addToast } from '~/stores/toast'
+import { getActionsCol } from '~/table/columns/action-col'
+import { Table } from '~/table/Table'
+import { CreateButton } from '~/ui/lib/CreateButton'
+import { EmptyMessage } from '~/ui/lib/EmptyMessage'
+import { TableActions, TableEmptyBox } from '~/ui/lib/Table'
+import { TipIcon } from '~/ui/lib/TipIcon'
+import { identityTypeLabel, roleColor } from '~/util/access'
+import { groupBy } from '~/util/array'
+import type * as PP from '~/util/path-params'
+
+const policyView = q(api.policyView, {})
+const projectPolicyView = ({ project }: PP.Project) =>
+  q(api.projectPolicyView, { path: { project } })
+const userList = q(api.userList, {})
+const groupList = q(api.groupList, {})
+
+const EmptyState = ({ onClick }: { onClick: () => void }) => (
+  <TableEmptyBox>
+    <EmptyMessage
+      icon={<Access24Icon />}
+      title="No authorized users"
+      body="Give permission to view, edit, or administer this project"
+      buttonText="Add user or group to project"
+      onClick={onClick}
+    />
+  </TableEmptyBox>
+)
+
+export async function clientLoader({ params }: LoaderFunctionArgs) {
+  const selector = getProjectSelector(params)
+  await Promise.all([
+    queryClient.prefetchQuery(policyView),
+    queryClient.prefetchQuery(projectPolicyView(selector)),
+    // used to resolve user names
+    queryClient.prefetchQuery(userList),
+    queryClient.prefetchQuery(groupList),
+  ])
+  return null
+}
+
+export const handle = titleCrumb('Roles')
+
+type UserRow = {
+  id: string
+  identityType: IdentityType
+  name: string
+  projectRole: RoleKey | undefined
+  roleBadges: { roleSource: string; roleName: RoleKey }[]
+}
+
+const colHelper = createColumnHelper<UserRow>()
+
+export default function ProjectAccessRolesTab() {
+  const [addModalOpen, setAddModalOpen] = useState(false)
+  const [editingUserRow, setEditingUserRow] = useState<UserRow | null>(null)
+  const projectSelector = useProjectSelector()
+
+  const { data: siloPolicy } = usePrefetchedQuery(policyView)
+  const siloRows = useUserRows(siloPolicy.roleAssignments, 'silo')
+
+  const { data: projectPolicy } = usePrefetchedQuery(projectPolicyView(projectSelector))
+  const projectRows = useUserRows(projectPolicy.roleAssignments, 'project')
+
+  const rows = useMemo(() => {
+    return groupBy(siloRows.concat(projectRows), (u) => u.id)
+      .map(([userId, userAssignments]) => {
+        const { name, identityType } = userAssignments[0]
+
+        const siloAccessRow = userAssignments.find((a) => a.roleSource === 'silo')
+        const projectAccessRow = userAssignments.find((a) => a.roleSource === 'project')
+
+        const roleBadges = R.sortBy(
+          [siloAccessRow, projectAccessRow].filter((r) => !!r),
+          (r) => roleOrder[r.roleName] // sorts strongest role first
+        )
+
+        return {
+          id: userId,
+          identityType,
+          name,
+          projectRole: projectAccessRow?.roleName,
+          roleBadges,
+        } satisfies UserRow
+      })
+      .sort(byGroupThenName)
+  }, [siloRows, projectRows])
+
+  const { mutateAsync: updatePolicy } = useApiMutation(api.projectPolicyUpdate, {
+    onSuccess: () => {
+      queryClient.invalidateEndpoint('projectPolicyView')
+      addToast({ content: 'Access removed' })
+    },
+    // TODO: handle 403
+  })
+
+  // TODO: checkboxes and bulk delete? not sure
+  // TODO: disable delete on permissions you can't delete
+
+  const columns = useMemo(
+    () => [
+      colHelper.accessor('name', { header: 'Name' }),
+      colHelper.accessor('identityType', {
+        header: 'Type',
+        cell: (info) => identityTypeLabel[info.getValue()],
+      }),
+      colHelper.accessor('roleBadges', {
+        header: () => (
+          <span className="inline-flex items-center">
+            Role
+            <TipIcon className="ml-2">
+              A user or group&apos;s effective role for this project is the strongest role
+              on either the silo or project
+            </TipIcon>
+          </span>
+        ),
+        cell: (info) => (
+          <ListPlusCell tooltipTitle="Other roles">
+            {info.getValue().map(({ roleName, roleSource }) => (
+              <Badge key={roleSource} color={roleColor[roleName]}>
+                {roleSource}.{roleName}
+              </Badge>
+            ))}
+          </ListPlusCell>
+        ),
+      }),
+
+      // TODO: tooltips on disabled elements explaining why
+      getActionsCol((row: UserRow) => [
+        {
+          label: 'Change role',
+          onActivate: () => setEditingUserRow(row),
+          disabled:
+            !row.projectRole && "You don't have permission to change this user's role",
+        },
+        // TODO: only show if you have permission to do this
+        {
+          label: 'Delete',
+          onActivate: confirmDelete({
+            doDelete: () =>
+              updatePolicy({
+                path: { project: projectSelector.project },
+                // we know policy is there, otherwise there's no row to display
+                body: deleteRole(row.id, projectPolicy),
+              }),
+            // TODO: explain that this will not affect the role inherited from
+            // the silo or roles inherited from group membership. Ideally we'd
+            // be able to say: this will cause the user to have an effective
+            // role of X. However we would have to look at their groups too.
+            label: (
+              <span>
+                the <HL>{row.projectRole}</HL> role for <HL>{row.name}</HL>
+              </span>
+            ),
+          }),
+          disabled: !row.projectRole && "You don't have permission to delete this user",
+        },
+      ]),
+    ],
+    [projectPolicy, projectSelector.project, updatePolicy]
+  )
+
+  const tableInstance = useReactTable({
+    columns,
+    data: rows,
+    getCoreRowModel: getCoreRowModel(),
+  })
+
+  return (
+    <>
+      <TableActions>
+        <CreateButton onClick={() => setAddModalOpen(true)}>Add user or group</CreateButton>
+      </TableActions>
+      {projectPolicy && addModalOpen && (
+        <ProjectAccessAddUserSideModal
+          onDismiss={() => setAddModalOpen(false)}
+          policy={projectPolicy}
+        />
+      )}
+      {projectPolicy && editingUserRow?.projectRole && (
+        <ProjectAccessEditUserSideModal
+          onDismiss={() => setEditingUserRow(null)}
+          policy={projectPolicy}
+          name={editingUserRow.name}
+          identityId={editingUserRow.id}
+          identityType={editingUserRow.identityType}
+          defaultValues={{ roleName: editingUserRow.projectRole }}
+        />
+      )}
+      {rows.length === 0 ? (
+        <EmptyState onClick={() => setAddModalOpen(true)} />
+      ) : (
+        <Table table={tableInstance} />
+      )}
+    </>
+  )
+}
