+++ app/routes.tsx
@@ -276,7 +276,18 @@ export const routes = createRoutesFromElements(
           />
         </Route>
 
-        <Route path="access" lazy={() => import('./pages/SiloAccessPage').then(convert)} />
+        <Route path="access" lazy={() => import('./pages/SiloAccessPage').then(convert)}>
+          <Route index element={<Navigate to="roles" replace />} />
+          <Route
+            path="roles"
+            lazy={() => import('./pages/SiloAccessRolesTab').then(convert)}
+          />
+          <Route path="users" lazy={() => import('./pages/AccessUsersTab').then(convert)} />
+          <Route
+            path="groups"
+            lazy={() => import('./pages/AccessGroupsTab').then(convert)}
+          />
+        </Route>
       </Route>
 
       {/* PROJECT */}
@@ -532,7 +543,23 @@ export const routes = createRoutesFromElements(
           <Route
             path="access"
             lazy={() => import('./pages/project/access/ProjectAccessPage').then(convert)}
-          />
+          >
+            <Route index element={<Navigate to="roles" replace />} />
+            <Route
+              path="roles"
+              lazy={() =>
+                import('./pages/project/access/ProjectAccessRolesTab').then(convert)
+              }
+            />
+            <Route
+              path="users"
+              lazy={() => import('./pages/AccessUsersTab').then(convert)}
+            />
+            <Route
+              path="groups"
+              lazy={() => import('./pages/AccessGroupsTab').then(convert)}
+            />
+          </Route>
           <Route
             lazy={() => import('./pages/project/affinity/AffinityPage').then(convert)}
             handle={{ crumb: 'Affinity Groups' }}
