+++ app/util/path-builder.spec.ts
@@ -73,6 +73,9 @@ test('path builder', () => {
         "profile": "/settings/profile",
         "project": "/projects/p/instances",
         "projectAccess": "/projects/p/access",
+        "projectAccessGroups": "/projects/p/access/groups",
+        "projectAccessRoles": "/projects/p/access/roles",
+        "projectAccessUsers": "/projects/p/access/users",
         "projectEdit": "/projects/p/edit",
         "projectImageEdit": "/projects/p/images/im/edit",
         "projectImages": "/projects/p/images",
@@ -83,6 +86,9 @@ test('path builder', () => {
         "serialConsole": "/projects/p/instances/i/serial-console",
         "silo": "/system/silos/s/idps",
         "siloAccess": "/access",
+        "siloAccessGroups": "/access/groups",
+        "siloAccessRoles": "/access/roles",
+        "siloAccessUsers": "/access/users",
         "siloFleetRoles": "/system/silos/s/fleet-roles",
         "siloIdps": "/system/silos/s/idps",
         "siloIdpsNew": "/system/silos/s/idps-new",
