+++ app/util/path-builder.ts
@@ -27,6 +27,9 @@ export const pb = {
   projectEdit: (params: PP.Project) => `${projectBase(params)}/edit`,
 
   projectAccess: (params: PP.Project) => `${projectBase(params)}/access`,
+  projectAccessRoles: (params: PP.Project) => `${projectBase(params)}/access/roles`,
+  projectAccessUsers: (params: PP.Project) => `${projectBase(params)}/access/users`,
+  projectAccessGroups: (params: PP.Project) => `${projectBase(params)}/access/groups`,
   projectImages: (params: PP.Project) => `${projectBase(params)}/images`,
   projectImagesNew: (params: PP.Project) => `${projectBase(params)}/images-new`,
   projectImageEdit: (params: PP.Image) =>
@@ -107,6 +110,9 @@ export const pb = {
 
   siloUtilization: () => '/utilization',
   siloAccess: () => '/access',
+  siloAccessRoles: () => '/access/roles',
+  siloAccessUsers: () => '/access/users',
+  siloAccessGroups: () => '/access/groups',
   siloImages: () => '/images',
   siloImageEdit: (params: PP.SiloImage) => `${pb.siloImages()}/${params.image}/edit`,
